microbit

Module gpio

Named GPIO pin types

This module maps the GPIO pin names as described in the Pins and Signals section of the micro:bit site Where appropriate the pins are restricted with the appropriate MODE from nrf-hal.

Structs

DisplayPins

GPIO pins connected to the LED matrix

Constants

NUM_COLS

NUM_ROWS

Type Aliases

BTN_A

BTN_B

COL1

COL2

COL3

COL4

COL5

COL6

COL7

COL8

COL9

EDGE00

EDGE01

EDGE02

EDGE03

EDGE04

EDGE05

EDGE06

EDGE07

EDGE08

EDGE09

EDGE10

EDGE11

EDGE12

EDGE13

EDGE14

EDGE15

EDGE16

EDGE19

EDGE20

MISO

MOSI

PAD0

PAD1

PAD2

ROW1

ROW2

ROW3

SCK

SCL

SDA

UART_RX

UART_TX