microbit::hal

Module clocks

Configuration and control of the High and Low Frequency Clock sources.

Structs

Clocks

A high level abstraction for the CLOCK peripheral.

ExternalOscillator

External Crystal Oscillator.

Internal

Internal/RC Oscillator.

LfOscStarted

Low Frequency Clock Started.

LfOscStopped

Low Frequency Clock Stopped.

LfOscSynthesized

Low Frequency Clock synthesize from High Frequency Clock.

Enums

LfOscConfiguration

Allowable configuration options for the low frequency oscillator when driven from an external crystal.

Constants

HFCLK_FREQ

LFCLK_FREQ

Low Frequency Clock Frequency (in Hz).