microbit::hal::gpio::p0

Struct P0_16

pub struct P0_16<MODE> { /* private fields */ }

Implementations

impl<MODE> P0_16<MODE>

pub fn into_floating_input(self) -> P0_16<Input<Floating>>

Convert the pin to be a floating input

pub fn into_pulldown_input(self) -> P0_16<Input<PullDown>>

pub fn into_pullup_input(self) -> P0_16<Input<PullUp>>

pub fn into_push_pull_output_drive( self, initial_output: Level, drive: DriveConfig, ) -> P0_16<Output<PushPull>>

Convert the pin to bepin a push-pull output with specified drive

pub fn into_push_pull_output( self, initial_output: Level, ) -> P0_16<Output<PushPull>>

Convert the pin to bepin a push-pull output with normal drive

pub fn into_open_drain_output( self, config: OpenDrainConfig, initial_output: Level, ) -> P0_16<Output<OpenDrain>>

Convert the pin to be an open-drain output

This method currently does not support configuring an internal pull-up or pull-down resistor.

pub fn into_open_drain_input_output( self, config: OpenDrainConfig, initial_output: Level, ) -> P0_16<Output<OpenDrainIO>>

Convert the pin to be an open-drain input/output

Similar to into_open_drain_output, but can also be read from.

This method currently does not support configuring an internal pull-up or pull-down resistor.

pub fn into_disconnected(self) -> P0_16<Disconnected>

Disconnects the pin.

In disconnected mode the pin cannot be used as input or output. It is primarily useful to reduce power usage.

pub fn degrade(self) -> Pin<MODE>

Degrade to a generic pin struct, which can be used with peripherals