microbit::hal

Module lpcomp

HAL interface for the LPCOMP peripheral.

In System ON, the LPCOMP can generate separate events on rising and falling edges of a signal, or sample the current state of the pin as being above or below the selected reference. The block can be configured to use any of the analog inputs on the device. Additionally, the low power comparator can be used as an analog wakeup source from System OFF. The comparator threshold can be programmed to a range of fractions of the supply voltage or to use an external analog reference input pin.

Structs

LpComp

A safe wrapper around the LPCOMP peripheral.

Enums

CompResult

Transition

VRef

Traits

LpCompInputPin

Trait to represent analog input pins.

LpCompRefPin

Trait to represent analog ref pins.