microbit::hal::pac::power

Module resetreas

Reset reason.

Structs

R

Register RESETREAS reader

RESETREAS_SPEC

Reset reason.

W

Register RESETREAS writer

Enums

DIF_A

Reset from wake-up from OFF mode detected by entering into debug interface mode.

DOG_A

Reset from watchdog detected.

LOCKUP_A

Reset from CPU lock-up detected.

LPCOMP_A

Reset from wake-up from OFF mode detected by the use of ANADETECT signal from LPCOMP.

OFF_A

Reset from wake-up from OFF mode detected by the use of DETECT signal from GPIO.

RESETPIN_A

Reset from pin-reset detected.

SREQ_A

Reset from AIRCR.SYSRESETREQ detected.

Type Aliases

DIF_R

Field DIF reader - Reset from wake-up from OFF mode detected by entering into debug interface mode.

DIF_W

Field DIF writer - Reset from wake-up from OFF mode detected by entering into debug interface mode.

DOG_R

Field DOG reader - Reset from watchdog detected.

DOG_W

Field DOG writer - Reset from watchdog detected.

LOCKUP_R

Field LOCKUP reader - Reset from CPU lock-up detected.

LOCKUP_W

Field LOCKUP writer - Reset from CPU lock-up detected.

LPCOMP_R

Field LPCOMP reader - Reset from wake-up from OFF mode detected by the use of ANADETECT signal from LPCOMP.

LPCOMP_W

Field LPCOMP writer - Reset from wake-up from OFF mode detected by the use of ANADETECT signal from LPCOMP.

OFF_R

Field OFF reader - Reset from wake-up from OFF mode detected by the use of DETECT signal from GPIO.

OFF_W

Field OFF writer - Reset from wake-up from OFF mode detected by the use of DETECT signal from GPIO.

RESETPIN_R

Field RESETPIN reader - Reset from pin-reset detected.

RESETPIN_W

Field RESETPIN writer - Reset from pin-reset detected.

SREQ_R

Field SREQ reader - Reset from AIRCR.SYSRESETREQ detected.

SREQ_W

Field SREQ writer - Reset from AIRCR.SYSRESETREQ detected.