microbit::hal::pac

Module radio

The radio.

Modules

base0

Radio base address 0. Decision point: START task.

base1

Radio base address 1. Decision point: START task.

bcc

Bit counter compare.

crccnf

CRC configuration.

crcinit

CRC initial value.

crcpoly

CRC polynomial.

crcstatus

CRC status of received packet.

dab

Device address base segment.

dacnf

Device address match configuration.

dai

Device address match index.

dap

Device address prefix.

datawhiteiv

Data whitening initial value.

events_address

Address event.

events_bcmatch

Bit counter reached bit count value specified in BCC register.

events_devmatch

A device address match occurred on the last received packet.

events_devmiss

No device address match occurred on the last received packet.

events_disabled

Disable event.

events_end

End event.

events_payload

Payload event.

events_ready

Ready event.

events_rssiend

Sampling of the receive signal strength complete. A new RSSI sample is ready for readout at the RSSISAMPLE register.

frequency

Frequency.

intenclr

Interrupt enable clear register.

intenset

Interrupt enable set register.

mode

Data rate and modulation.

override0

Trim value override register 0.

override1

Trim value override register 1.

override2

Trim value override register 2.

override3

Trim value override register 3.

override4

Trim value override register 4.

packetptr

Packet pointer. Decision point: START task.

pcnf0

Packet configuration 0.

pcnf1

Packet configuration 1.

power

Peripheral power control.

prefix0

Prefixes bytes for logical addresses 0 to 3.

prefix1

Prefixes bytes for logical addresses 4 to 7.

rssisample

RSSI sample.

rxaddresses

Receive address select.

rxcrc

Received CRC.

rxmatch

Received address.

shorts

Shortcuts for the radio.

state

Current radio state.

tasks_bcstart

Start the bit counter.

tasks_bcstop

Stop the bit counter.

tasks_disable

Disable radio.

tasks_rssistart

Start the RSSI and take one sample of the receive signal strength.

tasks_rssistop

Stop the RSSI measurement.

tasks_rxen

Enable radio in RX mode.

tasks_start

Start radio.

tasks_stop

Stop radio.

tasks_txen

Enable radio in TX mode.

test

Test features enable register.

tifs

Inter Frame Spacing in microseconds.

txaddress

Transmit address select.

txpower

Output power.

Structs

RegisterBlock

Register block

Type Aliases

BASE0

BASE0 (rw) register accessor: an alias for Reg<BASE0_SPEC>

BASE1

BASE1 (rw) register accessor: an alias for Reg<BASE1_SPEC>

BCC

BCC (rw) register accessor: an alias for Reg<BCC_SPEC>

CRCCNF

CRCCNF (rw) register accessor: an alias for Reg<CRCCNF_SPEC>

CRCINIT

CRCINIT (rw) register accessor: an alias for Reg<CRCINIT_SPEC>

CRCPOLY

CRCPOLY (rw) register accessor: an alias for Reg<CRCPOLY_SPEC>

CRCSTATUS

CRCSTATUS (r) register accessor: an alias for Reg<CRCSTATUS_SPEC>

DAB

DAB (rw) register accessor: an alias for Reg<DAB_SPEC>

DACNF

DACNF (rw) register accessor: an alias for Reg<DACNF_SPEC>

DAI

DAI (r) register accessor: an alias for Reg<DAI_SPEC>

DAP

DAP (rw) register accessor: an alias for Reg<DAP_SPEC>

DATAWHITEIV

DATAWHITEIV (rw) register accessor: an alias for Reg<DATAWHITEIV_SPEC>

EVENTS_ADDRESS

EVENTS_ADDRESS (rw) register accessor: an alias for Reg<EVENTS_ADDRESS_SPEC>

EVENTS_BCMATCH

EVENTS_BCMATCH (rw) register accessor: an alias for Reg<EVENTS_BCMATCH_SPEC>

EVENTS_DEVMATCH

EVENTS_DEVMATCH (rw) register accessor: an alias for Reg<EVENTS_DEVMATCH_SPEC>

EVENTS_DEVMISS

EVENTS_DEVMISS (rw) register accessor: an alias for Reg<EVENTS_DEVMISS_SPEC>

EVENTS_DISABLED

EVENTS_DISABLED (rw) register accessor: an alias for Reg<EVENTS_DISABLED_SPEC>

EVENTS_END

EVENTS_END (rw) register accessor: an alias for Reg<EVENTS_END_SPEC>

EVENTS_PAYLOAD

EVENTS_PAYLOAD (rw) register accessor: an alias for Reg<EVENTS_PAYLOAD_SPEC>

EVENTS_READY

EVENTS_READY (rw) register accessor: an alias for Reg<EVENTS_READY_SPEC>

EVENTS_RSSIEND

EVENTS_RSSIEND (rw) register accessor: an alias for Reg<EVENTS_RSSIEND_SPEC>

FREQUENCY

FREQUENCY (rw) register accessor: an alias for Reg<FREQUENCY_SPEC>

INTENCLR

INTENCLR (rw) register accessor: an alias for Reg<INTENCLR_SPEC>

INTENSET

INTENSET (rw) register accessor: an alias for Reg<INTENSET_SPEC>

MODE

MODE (rw) register accessor: an alias for Reg<MODE_SPEC>

OVERRIDE0

OVERRIDE0 (rw) register accessor: an alias for Reg<OVERRIDE0_SPEC>

OVERRIDE1

OVERRIDE1 (rw) register accessor: an alias for Reg<OVERRIDE1_SPEC>

OVERRIDE2

OVERRIDE2 (rw) register accessor: an alias for Reg<OVERRIDE2_SPEC>

OVERRIDE3

OVERRIDE3 (rw) register accessor: an alias for Reg<OVERRIDE3_SPEC>

OVERRIDE4

OVERRIDE4 (rw) register accessor: an alias for Reg<OVERRIDE4_SPEC>

PACKETPTR

PACKETPTR (rw) register accessor: an alias for Reg<PACKETPTR_SPEC>

PCNF0

PCNF0 (rw) register accessor: an alias for Reg<PCNF0_SPEC>

PCNF1

PCNF1 (rw) register accessor: an alias for Reg<PCNF1_SPEC>

POWER

POWER (rw) register accessor: an alias for Reg<POWER_SPEC>

PREFIX0

PREFIX0 (rw) register accessor: an alias for Reg<PREFIX0_SPEC>

PREFIX1

PREFIX1 (rw) register accessor: an alias for Reg<PREFIX1_SPEC>

RSSISAMPLE

RSSISAMPLE (r) register accessor: an alias for Reg<RSSISAMPLE_SPEC>

RXADDRESSES

RXADDRESSES (rw) register accessor: an alias for Reg<RXADDRESSES_SPEC>

RXCRC

RXCRC (r) register accessor: an alias for Reg<RXCRC_SPEC>

RXMATCH

RXMATCH (r) register accessor: an alias for Reg<RXMATCH_SPEC>

SHORTS

SHORTS (rw) register accessor: an alias for Reg<SHORTS_SPEC>

STATE

STATE (r) register accessor: an alias for Reg<STATE_SPEC>

TASKS_BCSTART

TASKS_BCSTART (w) register accessor: an alias for Reg<TASKS_BCSTART_SPEC>

TASKS_BCSTOP

TASKS_BCSTOP (w) register accessor: an alias for Reg<TASKS_BCSTOP_SPEC>

TASKS_DISABLE

TASKS_DISABLE (w) register accessor: an alias for Reg<TASKS_DISABLE_SPEC>

TASKS_RSSISTART

TASKS_RSSISTART (w) register accessor: an alias for Reg<TASKS_RSSISTART_SPEC>

TASKS_RSSISTOP

TASKS_RSSISTOP (w) register accessor: an alias for Reg<TASKS_RSSISTOP_SPEC>

TASKS_RXEN

TASKS_RXEN (w) register accessor: an alias for Reg<TASKS_RXEN_SPEC>

TASKS_START

TASKS_START (w) register accessor: an alias for Reg<TASKS_START_SPEC>

TASKS_STOP

TASKS_STOP (w) register accessor: an alias for Reg<TASKS_STOP_SPEC>

TASKS_TXEN

TASKS_TXEN (w) register accessor: an alias for Reg<TASKS_TXEN_SPEC>

TEST

TEST (rw) register accessor: an alias for Reg<TEST_SPEC>

TIFS

TIFS (rw) register accessor: an alias for Reg<TIFS_SPEC>

TXADDRESS

TXADDRESS (rw) register accessor: an alias for Reg<TXADDRESS_SPEC>

TXPOWER

TXPOWER (rw) register accessor: an alias for Reg<TXPOWER_SPEC>