microbit::hal::pac::radio

Struct RegisterBlock

pub struct RegisterBlock {

    pub tasks_txen: Reg<TASKS_TXEN_SPEC>,
    pub tasks_rxen: Reg<TASKS_RXEN_SPEC>,
    pub tasks_start: Reg<TASKS_START_SPEC>,
    pub tasks_stop: Reg<TASKS_STOP_SPEC>,
    pub tasks_disable: Reg<TASKS_DISABLE_SPEC>,
    pub tasks_rssistart: Reg<TASKS_RSSISTART_SPEC>,
    pub tasks_rssistop: Reg<TASKS_RSSISTOP_SPEC>,
    pub tasks_bcstart: Reg<TASKS_BCSTART_SPEC>,
    pub tasks_bcstop: Reg<TASKS_BCSTOP_SPEC>,
    pub events_ready: Reg<EVENTS_READY_SPEC>,
    pub events_address: Reg<EVENTS_ADDRESS_SPEC>,
    pub events_payload: Reg<EVENTS_PAYLOAD_SPEC>,
    pub events_end: Reg<EVENTS_END_SPEC>,
    pub events_disabled: Reg<EVENTS_DISABLED_SPEC>,
    pub events_devmatch: Reg<EVENTS_DEVMATCH_SPEC>,
    pub events_devmiss: Reg<EVENTS_DEVMISS_SPEC>,
    pub events_rssiend: Reg<EVENTS_RSSIEND_SPEC>,
    pub events_bcmatch: Reg<EVENTS_BCMATCH_SPEC>,
    pub shorts: Reg<SHORTS_SPEC>,
    pub intenset: Reg<INTENSET_SPEC>,
    pub intenclr: Reg<INTENCLR_SPEC>,
    pub crcstatus: Reg<CRCSTATUS_SPEC>,
    pub rxmatch: Reg<RXMATCH_SPEC>,
    pub rxcrc: Reg<RXCRC_SPEC>,
    pub dai: Reg<DAI_SPEC>,
    pub packetptr: Reg<PACKETPTR_SPEC>,
    pub frequency: Reg<FREQUENCY_SPEC>,
    pub txpower: Reg<TXPOWER_SPEC>,
    pub mode: Reg<MODE_SPEC>,
    pub pcnf0: Reg<PCNF0_SPEC>,
    pub pcnf1: Reg<PCNF1_SPEC>,
    pub base0: Reg<BASE0_SPEC>,
    pub base1: Reg<BASE1_SPEC>,
    pub prefix0: Reg<PREFIX0_SPEC>,
    pub prefix1: Reg<PREFIX1_SPEC>,
    pub txaddress: Reg<TXADDRESS_SPEC>,
    pub rxaddresses: Reg<RXADDRESSES_SPEC>,
    pub crccnf: Reg<CRCCNF_SPEC>,
    pub crcpoly: Reg<CRCPOLY_SPEC>,
    pub crcinit: Reg<CRCINIT_SPEC>,
    pub test: Reg<TEST_SPEC>,
    pub tifs: Reg<TIFS_SPEC>,
    pub rssisample: Reg<RSSISAMPLE_SPEC>,
    pub state: Reg<STATE_SPEC>,
    pub datawhiteiv: Reg<DATAWHITEIV_SPEC>,
    pub bcc: Reg<BCC_SPEC>,
    pub dab: [Reg<DAB_SPEC>; 8],
    pub dap: [Reg<DAP_SPEC>; 8],
    pub dacnf: Reg<DACNF_SPEC>,
    pub override0: Reg<OVERRIDE0_SPEC>,
    pub override1: Reg<OVERRIDE1_SPEC>,
    pub override2: Reg<OVERRIDE2_SPEC>,
    pub override3: Reg<OVERRIDE3_SPEC>,
    pub override4: Reg<OVERRIDE4_SPEC>,
    pub power: Reg<POWER_SPEC>,
    /* private fields */

}

Register block

Fields
tasks_txen: Reg<TASKS_TXEN_SPEC>

0x00 - Enable radio in TX mode.

tasks_rxen: Reg<TASKS_RXEN_SPEC>

0x04 - Enable radio in RX mode.

tasks_start: Reg<TASKS_START_SPEC>

0x08 - Start radio.

tasks_stop: Reg<TASKS_STOP_SPEC>

0x0c - Stop radio.

tasks_disable: Reg<TASKS_DISABLE_SPEC>

0x10 - Disable radio.

tasks_rssistart: Reg<TASKS_RSSISTART_SPEC>

0x14 - Start the RSSI and take one sample of the receive signal strength.

tasks_rssistop: Reg<TASKS_RSSISTOP_SPEC>

0x18 - Stop the RSSI measurement.

tasks_bcstart: Reg<TASKS_BCSTART_SPEC>

0x1c - Start the bit counter.

tasks_bcstop: Reg<TASKS_BCSTOP_SPEC>

0x20 - Stop the bit counter.

events_ready: Reg<EVENTS_READY_SPEC>

0x100 - Ready event.

events_address: Reg<EVENTS_ADDRESS_SPEC>

0x104 - Address event.

events_payload: Reg<EVENTS_PAYLOAD_SPEC>

0x108 - Payload event.

events_end: Reg<EVENTS_END_SPEC>

0x10c - End event.

events_disabled: Reg<EVENTS_DISABLED_SPEC>

0x110 - Disable event.

events_devmatch: Reg<EVENTS_DEVMATCH_SPEC>

0x114 - A device address match occurred on the last received packet.

events_devmiss: Reg<EVENTS_DEVMISS_SPEC>

0x118 - No device address match occurred on the last received packet.

events_rssiend: Reg<EVENTS_RSSIEND_SPEC>

0x11c - Sampling of the receive signal strength complete. A new RSSI sample is ready for readout at the RSSISAMPLE register.

events_bcmatch: Reg<EVENTS_BCMATCH_SPEC>

0x128 - Bit counter reached bit count value specified in BCC register.

shorts: Reg<SHORTS_SPEC>

0x200 - Shortcuts for the radio.

intenset: Reg<INTENSET_SPEC>

0x304 - Interrupt enable set register.

intenclr: Reg<INTENCLR_SPEC>

0x308 - Interrupt enable clear register.

crcstatus: Reg<CRCSTATUS_SPEC>

0x400 - CRC status of received packet.

rxmatch: Reg<RXMATCH_SPEC>

0x408 - Received address.

rxcrc: Reg<RXCRC_SPEC>

0x40c - Received CRC.

dai: Reg<DAI_SPEC>

0x410 - Device address match index.

packetptr: Reg<PACKETPTR_SPEC>

0x504 - Packet pointer. Decision point: START task.

frequency: Reg<FREQUENCY_SPEC>

0x508 - Frequency.

txpower: Reg<TXPOWER_SPEC>

0x50c - Output power.

mode: Reg<MODE_SPEC>

0x510 - Data rate and modulation.

pcnf0: Reg<PCNF0_SPEC>

0x514 - Packet configuration 0.

pcnf1: Reg<PCNF1_SPEC>

0x518 - Packet configuration 1.

base0: Reg<BASE0_SPEC>

0x51c - Radio base address 0. Decision point: START task.

base1: Reg<BASE1_SPEC>

0x520 - Radio base address 1. Decision point: START task.

prefix0: Reg<PREFIX0_SPEC>

0x524 - Prefixes bytes for logical addresses 0 to 3.

prefix1: Reg<PREFIX1_SPEC>

0x528 - Prefixes bytes for logical addresses 4 to 7.

txaddress: Reg<TXADDRESS_SPEC>

0x52c - Transmit address select.

rxaddresses: Reg<RXADDRESSES_SPEC>

0x530 - Receive address select.

crccnf: Reg<CRCCNF_SPEC>

0x534 - CRC configuration.

crcpoly: Reg<CRCPOLY_SPEC>

0x538 - CRC polynomial.

crcinit: Reg<CRCINIT_SPEC>

0x53c - CRC initial value.

test: Reg<TEST_SPEC>

0x540 - Test features enable register.

tifs: Reg<TIFS_SPEC>

0x544 - Inter Frame Spacing in microseconds.

rssisample: Reg<RSSISAMPLE_SPEC>

0x548 - RSSI sample.

state: Reg<STATE_SPEC>

0x550 - Current radio state.

datawhiteiv: Reg<DATAWHITEIV_SPEC>

0x554 - Data whitening initial value.

bcc: Reg<BCC_SPEC>

0x560 - Bit counter compare.

dab: [Reg<DAB_SPEC>; 8]

0x600..0x620 - Device address base segment.

dap: [Reg<DAP_SPEC>; 8]

0x620..0x640 - Device address prefix.

dacnf: Reg<DACNF_SPEC>

0x640 - Device address match configuration.

override0: Reg<OVERRIDE0_SPEC>

0x724 - Trim value override register 0.

override1: Reg<OVERRIDE1_SPEC>

0x728 - Trim value override register 1.

override2: Reg<OVERRIDE2_SPEC>

0x72c - Trim value override register 2.

override3: Reg<OVERRIDE3_SPEC>

0x730 - Trim value override register 3.

override4: Reg<OVERRIDE4_SPEC>

0x734 - Trim value override register 4.

power: Reg<POWER_SPEC>

0xffc - Peripheral power control.