microbit::hal::pac::rtc0

Module evtenclr

Disable events routing to PPI. The reading of this register gives the value of EVTEN.

Structs

EVTENCLR_SPEC

Disable events routing to PPI. The reading of this register gives the value of EVTEN.

R

Register EVTENCLR reader

W

Register EVTENCLR writer

Enums

COMPARE0_A

Disable routing to PPI of COMPARE[0] event.

COMPARE0_AW

Disable routing to PPI of COMPARE[0] event.

COMPARE1_A

Disable routing to PPI of COMPARE[1] event.

COMPARE1_AW

Disable routing to PPI of COMPARE[1] event.

COMPARE2_A

Disable routing to PPI of COMPARE[2] event.

COMPARE2_AW

Disable routing to PPI of COMPARE[2] event.

COMPARE3_A

Disable routing to PPI of COMPARE[3] event.

COMPARE3_AW

Disable routing to PPI of COMPARE[3] event.

OVRFLW_A

Disable routing to PPI of OVRFLW event.

OVRFLW_AW

Disable routing to PPI of OVRFLW event.

TICK_A

Disable routing to PPI of TICK event.

TICK_AW

Disable routing to PPI of TICK event.

Type Aliases

COMPARE0_R

Field COMPARE0 reader - Disable routing to PPI of COMPARE[0] event.

COMPARE0_W

Field COMPARE0 writer - Disable routing to PPI of COMPARE[0] event.

COMPARE1_R

Field COMPARE1 reader - Disable routing to PPI of COMPARE[1] event.

COMPARE1_W

Field COMPARE1 writer - Disable routing to PPI of COMPARE[1] event.

COMPARE2_R

Field COMPARE2 reader - Disable routing to PPI of COMPARE[2] event.

COMPARE2_W

Field COMPARE2 writer - Disable routing to PPI of COMPARE[2] event.

COMPARE3_R

Field COMPARE3 reader - Disable routing to PPI of COMPARE[3] event.

COMPARE3_W

Field COMPARE3 writer - Disable routing to PPI of COMPARE[3] event.

OVRFLW_R

Field OVRFLW reader - Disable routing to PPI of OVRFLW event.

OVRFLW_W

Field OVRFLW writer - Disable routing to PPI of OVRFLW event.

TICK_R

Field TICK reader - Disable routing to PPI of TICK event.

TICK_W

Field TICK writer - Disable routing to PPI of TICK event.